Factorize whitelist methods
